refactor: centralize shared UI components
