--- src/components/layout/index.ts
@@ -1,1 +1,0 @@
-export { default as Navbar } from './Navbar';
