+++ src/components/shared/Card.test.tsx
@@ -1,7 +1,15 @@
 import { render, screen } from '@testing-library/react';
-import { Card } from './Card';
+import { Card, CardHeader, CardTitle, CardContent } from './Card';
 
 test('renders card children', () => {
-  render(<Card>inside</Card>);
+  render(
+    <Card>
+      <CardHeader>
+        <CardTitle>Title</CardTitle>
+      </CardHeader>
+      <CardContent>inside</CardContent>
+    </Card>
+  );
+  expect(screen.getByText('Title')).toBeInTheDocument();
   expect(screen.getByText('inside')).toBeInTheDocument();
 });
