+++ src/components/shared/Card.tsx
@@ -5,16 +5,22 @@ interface Props {
   className?: string;
 }
 
-export const Card: React.FC<Props> = ({ children, className = '' }) => {
-  return (
-    <div className={`
-      bg-white dark:bg-gray-800 
-      rounded-lg shadow-sm 
-      border border-gray-200 dark:border-gray-700
-      p-6
-      ${className}
-    `}>
-      {children}
-    </div>
-  );
-};
+export const Card: React.FC<Props> = ({ children, className = '' }) => (
+  <div
+    className={`bg-white dark:bg-gray-800 rounded-lg shadow-sm border border-gray-200 dark:border-gray-700 p-6 ${className}`}
+  >
+    {children}
+  </div>
+);
+
+export const CardHeader: React.FC<Props> = ({ children, className = '' }) => (
+  <div className={`px-6 py-4 border-b border-gray-200 dark:border-gray-700 ${className}`}>{children}</div>
+);
+
+export const CardTitle: React.FC<Props> = ({ children, className = '' }) => (
+  <h3 className={`text-lg font-semibold ${className}`}>{children}</h3>
+);
+
+export const CardContent: React.FC<Props> = ({ children, className = '' }) => (
+  <div className={`px-6 py-4 ${className}`}>{children}</div>
+);
