+++ src/components/shared/Navbar.tsx
@@ -77,3 +77,4 @@ const Navbar: React.FC = () => {
 };
 
 export default Navbar;
+export { Navbar };
