+++ src/components/shared/Separator.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import { Separator } from './separator';
+import { Separator } from './Separator';
 
 test('renders separator element', () => {
   const { container } = render(<Separator />);
