+++ src/components/shared/index.ts
@@ -1,0 +1,12 @@
+export { Alert, AlertDescription } from './Alert';
+export { Badge } from './Badge';
+export { Button } from './Button';
+export { Card, CardHeader, CardTitle, CardContent } from './Card';
+export { Checkbox } from './Checkbox';
+export { ErrorBoundary } from './ErrorBoundary';
+export { Input } from './Input';
+export { Navbar } from './Navbar';
+export { ProgressBar } from './ProgressBar';
+export { Select } from './Select';
+export { Separator } from './Separator';
+export { Tabs, TabsList, TabsTrigger, TabsContent } from './Tabs';
